Reorganize the tests
The test suite had grown to many hidden interdependencies
and used a hacky workaround to simulate changing the workspace.

This change reorganizes the tests into several suites,
some running with the main test workspace
and others with the subdirectory workspace,
and runs them in separate instances of vscode
so their extension hosts don't accumulate too much state.

The vscode launch configuration runs all tests for the main test workspace.
We could add more configurations for the smaller suites and subdirectory workspace
but that doesn't seem worth the effort right now.
And for debugging we can always `.only` a failing test.