Unset environment variables in the extension host
Some programs insist on distinguishing null and unset variables.
We still can't unset variables in the terminal
but we *can* do so in the extension host (process.env).

This change explicitly handles null/undefined
in the cached environment and in the extension host.

Regarding null/undefined:
direnv sends us JSON
and so uses null for unset variables
because unmodified variables are implicltly undefined.
Node process.env is a JavaScript object
so unset variables are undefined.
We use null for the data exported by direnv
and undefined everywhere else.

Previously:
79a4841ea6700cedcaa7849126c1acf0b5726e92